+++ x/intrarelayer/client/tx.go
@@ -28,6 +28,7 @@ func NewTxCmd() *cobra.Command {
 
 	txCmd.AddCommand(
 		NewConvertCoinCmd(),
+		NewConvertERC20Cmd(),
 	)
 	return txCmd
 }