+++ CHANGELOG.md
@@ -35,7 +35,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
-## Unreleased
+## [v0.3.0] - 2021-11-24
 
 ## API Breaking